% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A singleton numeric vector for which is interesting to plot.}

\item{year}{A character string or numeric vector which contains the year of
your interest to plot}
}
\value{
\code{NULL}
}
\description{
This function plots the places were did it happen the incidentes recorded in
the year \code{year} in the state \code{state.num}. The function verify that
state is correct and that the state has incidents reported in the year.
}
\examples{
fars_map_state(1, 2013)
fars_map_state(4, 2014)

}
