% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path of the file of the subject
to be read}
}
\value{
This function returns a dataset of class \code{tbl_df} (see \code{dplyr}
package) which is locate in \code{filename} file.
}
\description{
This function reads a path from a csv file and reads the file by returning
the dataset as a \code{tbl_df} object of the \code{dplyr} package. In case
the path does not exist the function throws a message
}
\examples{
tf <- tempfile()
fars_read(tf)

}
\seealso{
tbl_df
}
