% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of character string or numeric vector which contains the years
of your interest.}
}
\value{
This function returns a list of datasets of class \code{\link{tbl_df}}
with columns 'MONTH' and 'year'.
}
\description{
This function reads a list of files (each one with the structure required by
the function \code{fras_read}) and return the subset of columns 'MONTH' and
'year' where 'year' is tthe year in \code{years} param. The function throws a error
if any number in \code{years} is not finded.
}
\examples{
fars_read_years(c('2013', '2014', '2015'))
fars_read_years(2013:2015)

}
\seealso{
\code{\link{tbl_df}}
}
