% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of character string or numeric vector which contains the years
of your interest.}
}
\value{
This function returns a dataset of class \code{\link{tbl_df}}
that summarises the contents of files of years within the list \code{years} by
year and month and which the columns are the years and the month.
}
\description{
This function summarise the counts of observations by year and month
from the files contained in the list \code{years} param.
}
\examples{
fars_summarize_years(c('2013', '2014', '2015'))
fars_summarize_years(2013:2015)

}
\seealso{
\code{\link{tbl_df}}
}
