% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string or numeric vector which contains the year of
your interest.}
}
\value{
This function returns a character vector, a.k.a \code{string} in other
programming languages, with the structure mentioned in the function description.
}
\description{
This function builds a string with the following structure:

\code{accident_\strong{year}.csv.bz2}
}
\examples{
make_filename('2013')
make_filename(2014)

}
